Add argument to monitor external sets during training
Add an extra argument to the train function that allows to monitor the
accuracy on external validation sets during training.
In the `train_model.py` script, these extra datasets can be indicated
using the flag "--monitor". Every `print_every` steps the accuracy on
all monitor sets is computed and printed.
Defaulut behaviour remains the same: the dev set is used for model
saving and tuning the optimizer. The monitor sets are only used for
monitoring and do not interfere with the training.